requirements.txt
opencv-python==3.4.0
h5py==2.9.0
numpy==1.17.0
tqdm==4.19.5
matplotlib==2.2.3
seaborn==0.9.0
scipy==1.4.1
scikit-learn==0.19.1
tensorflow-gpu==1.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0